added Realtime service

+++ backend/Common/ProjectX.Core/Auth/ICurrentUser.cs
@@ -2,6 +2,6 @@
 
 public interface ICurrentUser
 {
-    long IdentityId { get; }
+    int IdentityId { get; }
     string IdentityRole { get; }
 }
